Encapsulate ResolvedPathDb and router validation. Remove behaviour from ComponentDb.
