+++ libs/pavex/src/compiler/analyses/computations.rs
@@ -1,19 +1,9 @@
 use ahash::HashMap;
-use guppy::graph::PackageGraph;
-use miette::{miette, NamedSource};
-use rustdoc_types::ItemEnum;
-use syn::spanned::Spanned;
 
-use crate::compiler::analyses::raw_user_components::{RawUserComponentDb, RawUserComponentId};
-use crate::compiler::analyses::resolved_paths::ResolvedPathDb;
+use crate::compiler::analyses::raw_user_components::RawUserComponentId;
 use crate::compiler::computation::Computation;
 use crate::compiler::interner::Interner;
 use crate::compiler::resolvers::{resolve_callable, CallableResolutionError};
-use crate::diagnostic;
-use crate::diagnostic::{
-    convert_proc_macro_span, convert_rustdoc_span, AnnotatedSnippet, CompilerDiagnostic,
-    LocationExt, SourceSpanExt,
-};
 use crate::language::{Callable, ResolvedPath};
 use crate::rustdoc::CrateCollection;
 
@@ -26,246 +16,37 @@ pub(crate) struct ComputationDb {
 }
 
 impl ComputationDb {
-    pub fn build(
-        component_db: &RawUserComponentDb,
-        resolved_path_db: &ResolvedPathDb,
-        package_graph: &PackageGraph,
-        krate_collection: &CrateCollection,
-        diagnostics: &mut Vec<miette::Error>,
-    ) -> Self {
-        let mut self_ = Self {
+    /// Initialize a new (empty) computation database.
+    pub fn new() -> Self {
+        Self {
             interner: Interner::new(),
             component_id2callable_id: Default::default(),
-        };
-        for (component_id, _) in component_db.iter() {
-            let resolved_path = &resolved_path_db[component_id];
-            if let Err(e) =
-                self_.resolve_callable(krate_collection, resolved_path, Some(component_id))
-            {
-                Self::capture_diagnostics(
-                    e,
-                    component_id,
-                    component_db,
-                    package_graph,
-                    diagnostics,
-                );
-            }
         }
-        self_
     }
 
-    pub(crate) fn resolve_callable(
+    /// Try to resolve a callable from a resolved path.
+    /// Returns the callable's id in the interner if it succeeds, an error otherwise.
+    pub(crate) fn resolve_and_intern(
         &mut self,
         krate_collection: &CrateCollection,
         resolved_path: &ResolvedPath,
-        component_id: Option<RawUserComponentId>,
+        raw_user_component_id: Option<RawUserComponentId>,
     ) -> Result<ComputationId, CallableResolutionError> {
         let callable = resolve_callable(krate_collection, resolved_path)?;
         let callable_id = self.interner.get_or_intern(callable.into());
-        if let Some(component_id) = component_id {
+        if let Some(raw_user_id) = raw_user_component_id {
             self.component_id2callable_id
-                .insert(component_id, callable_id);
+                .insert(raw_user_id, callable_id);
         }
         Ok(callable_id)
     }
 
+    /// Retrieve the id for a computation from the interner, or insert it if it doesn't exist.
     pub(crate) fn get_or_intern(
         &mut self,
         computation: impl Into<Computation<'static>>,
     ) -> ComputationId {
         self.interner.get_or_intern(computation.into())
-    }
-
-    fn capture_diagnostics(
-        e: CallableResolutionError,
-        component_id: RawUserComponentId,
-        component_db: &RawUserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = component_db.get_location(component_id);
-        let component = &component_db[component_id];
-        let callable_type = component.callable_type();
-        let source = match location.source_file(package_graph) {
-            Ok(source) => source,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        match e {
-            CallableResolutionError::UnknownCallable(_) => {
-                let label = diagnostic::get_f_macro_invocation_span(&source, location)
-                    .map(|s| s.labeled(format!("The {callable_type} that we can't resolve")));
-                let diagnostic = CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .help("Check that the path is spelled correctly and that the function (or method) is marked as `pub`.".into())
-                    .build();
-                diagnostics.push(diagnostic.into());
-            }
-            CallableResolutionError::ParameterResolutionError(ref inner_error) => {
-                let definition_snippet = {
-                    if let Some(definition_span) = &inner_error.callable_item.span {
-                        match diagnostic::read_source_file(
-                            &definition_span.filename,
-                            &package_graph.workspace(),
-                        ) {
-                            Ok(source_contents) => {
-                                let span = convert_rustdoc_span(
-                                    &source_contents,
-                                    definition_span.to_owned(),
-                                );
-                                let span_contents =
-                                    &source_contents[span.offset()..(span.offset() + span.len())];
-                                let input = match &inner_error.callable_item.inner {
-                                    ItemEnum::Function(_) => {
-                                        if let Ok(item) = syn::parse_str::<syn::ItemFn>(span_contents) {
-                                            let mut inputs = item.sig.inputs.iter();
-                                            inputs.nth(inner_error.parameter_index).cloned()
-                                        } else if let Ok(item) =
-                                            syn::parse_str::<syn::ImplItemMethod>(span_contents)
-                                        {
-                                            let mut inputs = item.sig.inputs.iter();
-                                            inputs.nth(inner_error.parameter_index).cloned()
-                                        } else {
-                                            panic!(
-                                                "Could not parse as a function or method:\n{span_contents}"
-                                            )
-                                        }
-                                    }
-                                    _ => unreachable!(),
-                                }
-                                    .unwrap();
-                                let s = convert_proc_macro_span(
-                                    span_contents,
-                                    match input {
-                                        syn::FnArg::Typed(typed) => typed.ty.span(),
-                                        syn::FnArg::Receiver(r) => r.span(),
-                                    },
-                                );
-                                let label = miette::SourceSpan::new(
-                                    // We must shift the offset forward because it's the
-                                    // offset from the beginning of the file slice that
-                                    // we deserialized, instead of the entire file
-                                    (s.offset() + span.offset()).into(),
-                                    s.len().into(),
-                                )
-                                .labeled("I don't know how handle this parameter".into());
-                                let source_path = definition_span.filename.to_str().unwrap();
-                                Some(AnnotatedSnippet::new(
-                                    NamedSource::new(source_path, source_contents),
-                                    label,
-                                ))
-                            }
-                            Err(e) => {
-                                tracing::warn!("Could not read source file: {:?}", e);
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    }
-                };
-                let label = diagnostic::get_f_macro_invocation_span(&source, location)
-                    .map(|s| s.labeled(format!("The {callable_type} was registered here")));
-                let diagnostic = CompilerDiagnostic::builder(source, e.clone())
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .build();
-                diagnostics.push(diagnostic.into());
-            }
-            CallableResolutionError::UnsupportedCallableKind(ref inner_error) => {
-                let label = diagnostic::get_f_macro_invocation_span(&source, location)
-                    .map(|s| s.labeled(format!("It was registered as a {callable_type} here")));
-                let message = format!("I can work with functions and static methods, but `{}` is neither.\nIt is {} and I don't know how to use it as a {}.", inner_error.import_path, inner_error.item_kind, callable_type);
-                let error = anyhow::anyhow!(e).context(message);
-                diagnostics.push(
-                    CompilerDiagnostic::builder(source, error)
-                        .optional_label(label)
-                        .build()
-                        .into(),
-                );
-            }
-            CallableResolutionError::OutputTypeResolutionError(ref inner_error) => {
-                let annotated_snippet = {
-                    if let Some(definition_span) = &inner_error.callable_item.span {
-                        match diagnostic::read_source_file(
-                            &definition_span.filename,
-                            &package_graph.workspace(),
-                        ) {
-                            Ok(source_contents) => {
-                                let span = convert_rustdoc_span(
-                                    &source_contents,
-                                    definition_span.to_owned(),
-                                );
-                                let span_contents = source_contents
-                                    [span.offset()..(span.offset() + span.len())]
-                                    .to_string();
-                                let output = match &inner_error.callable_item.inner {
-                                    ItemEnum::Function(_) => {
-                                        if let Ok(item) =
-                                            syn::parse_str::<syn::ItemFn>(&span_contents)
-                                        {
-                                            item.sig.output
-                                        } else if let Ok(item) =
-                                            syn::parse_str::<syn::ImplItemMethod>(&span_contents)
-                                        {
-                                            item.sig.output
-                                        } else {
-                                            panic!(
-                                                "Could not parse as a function or method:\n{span_contents}"
-                                            )
-                                        }
-                                    }
-                                    _ => unreachable!(),
-                                };
-                                match output {
-                                    syn::ReturnType::Default => None,
-                                    syn::ReturnType::Type(_, type_) => Some(type_.span()),
-                                }
-                                .map(|s| {
-                                    let s = convert_proc_macro_span(&span_contents, s);
-                                    let label = miette::SourceSpan::new(
-                                        // We must shift the offset forward because it's the
-                                        // offset from the beginning of the file slice that
-                                        // we deserialized, instead of the entire file
-                                        (s.offset() + span.offset()).into(),
-                                        s.len().into(),
-                                    )
-                                    .labeled("The output type that I can't handle".into());
-                                    AnnotatedSnippet::new(
-                                        NamedSource::new(
-                                            definition_span.filename.to_str().unwrap(),
-                                            source_contents,
-                                        ),
-                                        label,
-                                    )
-                                })
-                            }
-                            Err(e) => {
-                                tracing::warn!("Could not read source file: {:?}", e);
-                                None
-                            }
-                        }
-                    } else {
-                        None
-                    }
-                };
-
-                let label = diagnostic::get_f_macro_invocation_span(&source, location)
-                    .map(|s| s.labeled(format!("The {callable_type} was registered here")));
-                diagnostics.push(
-                    CompilerDiagnostic::builder(source, e.clone())
-                        .optional_label(label)
-                        .optional_additional_annotated_snippet(annotated_snippet)
-                        .build()
-                        .into(),
-                )
-            }
-            CallableResolutionError::CannotGetCrateData(_) => {
-                diagnostics.push(miette!(e));
-            }
-        }
     }
 }
